Move backend/lib/sfx/ to backend/sfx/
