+++ backend/sfx/testdata/sfx-context-object-valid.xml
